Update validateScore method to accept Double type for score parameter

+++ Server/src/main/java/com/grade/forge/grading/service/SubmissionGradeService.java
@@ -213,7 +213,7 @@ public class SubmissionGradeService {
         }
     }
 
-    private void validateScore(RubricSubCriteria subCriteria, Integer score) {
+    private void validateScore(RubricSubCriteria subCriteria, Double score) {
         if (score < 0) {
             throw new IllegalArgumentException("awardedScore cannot be negative");
         }
